backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
google-generativeai==0.7.2
pydantic>=2.0
